requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
jinja2==3.1.6
python-multipart==0.0.20
pypdf==6.7.5
reportlab==4.4.1
Pillow==11.2.1
